backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Task Queue
celery==5.3.6
redis==5.0.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data Processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2
scipy==1.12.0
matplotlib==3.8.2
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1